requirements.txt
absl_py==1.4.0+computecanada
aiohttp==3.8.4+computecanada
aiosignal==1.3.1+computecanada
albumentations==1.3.0+computecanada
antlr4_python3_runtime==4.9.3+computecanada
async_timeout==4.0.2+computecanada
attrs==22.2.0+computecanada
awscli==1.27.113
awscli-plugin-endpoint==0.4
botocore==1.29.113
cachetools==5.3.0+computecanada
certifi==2022.12.7+computecanada
charset_normalizer==3.1.0+computecanada
colorama==0.4.4+computecanada
docutils==0.16+computecanada
efficientnet-pytorch==0.6.3
frozenlist==1.3.3+computecanada
fsspec==2023.4.0+computecanada
google_auth==2.16.1+computecanada
google_auth_oauthlib==1.0.0+computecanada
grpcio==1.51.3+computecanada
idna==3.4+computecanada
imagecodecs==2023.1.23+computecanada
imageio==2.27.0+computecanada
jmespath==1.0.1+computecanada
joblib==1.2.0+computecanada
lightning_utilities==0.8.0+computecanada
llvmlite==0.39.1+computecanada
Markdown==3.4.1+computecanada
multidict==6.0.4+computecanada
munch==2.5.0+computecanada
networkx==3.1+computecanada
numba==0.56.3+computecanada
numpy==1.23.2+computecanada
oauthlib==3.2.2+computecanada
omegaconf==2.3.0+computecanada
opencv==1.0.1+computecanada
overrides==3.1.0+computecanada
packaging==23.0+computecanada
Pillow_SIMD==9.0.0.post1+computecanada
pretrainedmodels==0.7.4+computecanada
protobuf==3.20.3+computecanada
pyasn1==0.4.8+computecanada
pyasn1-modules==0.2.8+computecanada
python_dotenv==0.20.0+computecanada
pytorch_lightning==1.9.1+computecanada
PyWavelets==1.3.0+computecanada
PyYAML==5.4.1+computecanada
qudida==0.0.4+computecanada
requests==2.28.2+computecanada
requests_oauthlib==1.3.1+computecanada
rsa==4.7.2+computecanada
s3transfer==0.6.0+computecanada
scikit_image==0.19.3+computecanada
scikit_learn==1.2.1+computecanada
segmentation-models-pytorch @ file:///home/mahzare/seg/segmentation_models.pytorch
shapely==2.0.1+computecanada
tensorboard==2.12.1+computecanada
tensorboard_data_server==0.7.0+computecanada
tensorboard_plugin_wit==1.8.1+computecanada
tensorboardX==2.6+computecanada
threadpoolctl==3.1.0+computecanada
tifffile==2023.3.21+computecanada
timm==0.4.12+computecanada
torch==1.13.1+computecanada
torchmetrics==0.11.1+computecanada
torchvision==0.14.1+computecanada
tqdm==4.65.0+computecanada
typing_extensions==4.5.0+computecanada
urllib3==1.26.15+computecanada
Werkzeug==2.2.3+computecanada
yarl==1.8.2+computecanada